feat: release v2.2.0 with parallelization and memory optimizations
- Deleted outdated `top_10_closest_targets.csv` to streamline repository.
- Upgraded pipeline to v2.2.0 with parallel supplemental data processing using `ProcessPoolExecutor` for faster dataset aggregation.
- Optimized memory usage with intelligent type casting (`float32`, `int32`) and in-place operations during preprocessing.
- Updated `README` with new features, including parallelization, memory-efficient I/O, and documentation enhancements.
- Reflected changes in CatBoost training data with additional metrics and iterations for detailed analysis.

+++ src/orchestrator.py
@@ -88,7 +88,7 @@ def run_pipeline(data_dir=None, folds=3, prefer_lightgbm=True, custom_out=None, 
     test_df = fix_known_anomalies(test_df)
     test_df = add_engineered_features(test_df)
     
-    y_temp = train_df[config.target_col].astype(int)
+    y_temp = train_df[config.target_col].astype(np.int8)
     x_temp = train_df.drop(columns=[config.target_col])
     
     # Use a sample for speed
@@ -125,7 +125,7 @@ def run_pipeline(data_dir=None, folds=3, prefer_lightgbm=True, custom_out=None, 
     if id_col not in train_df.columns or id_col not in test_df.columns:
         raise ValueError(f"Expected {id_col} column in both train and test.")
 
-    y = train_df[target_col].astype(int)
+    y = train_df[target_col].astype(np.int8)
     test_ids = test_df[id_col]
 
     x_train = train_df.drop(columns=[target_col])
